backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
faiss-cpu
openai
python-dotenv
pydantic
python-multipart
langgraph>=0.1.7
rank-bm25>=0.2.2
pypdf>=3.9.0
python-docx>=1.1.0
trafilatura>=1.6.3
tiktoken>=0.7.0
cohere>=5.5.8
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.32
pgvector>=0.2.5
